pull database ops out
