+++ src/MdGraph/App/RunCommand.hs
@@ -1,3 +1,4 @@
+{-#  LANGUAGE RankNTypes  #-}
 {-# LANGUAGE  NamedFieldPuns  #-}
 module MdGraph.App.RunCommand where
 
@@ -5,6 +6,7 @@ import           Aux.HashSet
 import           MdGraph.TagDirection
 
 import           Control.Applicative            ( Applicative(liftA2) )
+import           Control.Exception              ( throwIO )
 import           Control.Monad                  ( join )
 import           Control.Monad.Except           ( MonadError(throwError)
                                                 , MonadIO(liftIO)
@@ -20,15 +22,19 @@ import           MdGraph.App                    ( App
                                                 , Env(config)
                                                 )
 import           MdGraph.App.Command
-import           MdGraph.App.Logger             ( logDebug
+import           MdGraph.App.Logger             ( Logs
+                                                , logDebug
                                                 , logInfo
                                                 )
 import           MdGraph.Config                 ( Config
                                                   ( dbConnString
                                                   , libraryPath
                                                   )
+                                                , HasConfig(getConfig)
                                                 )
-import           MdGraph.File                   ( trueAbsolutePath )
+import           MdGraph.File                   ( Files
+                                                , trueAbsolutePath
+                                                )
 import           MdGraph.Persist.Query         as Q
 import           MdGraph.Persist.Schema
 import qualified MdGraph.Persist.Schema        as Edge
@@ -48,70 +54,71 @@ runCommand (Backlinks options) = runBacklinks options
 runCommand Statics             = throwError "NYI"
 runCommand Populate            = pure mempty
 
-runOrphans :: App [Document]
+runOrphans :: (Monad m, RunsQuery m, Logs m) => m [Document]
 runOrphans = do
-  connString <- asks $ dbConnString . config
-  orphanDocs <- Q.orphans connString
-  let orphanCount = length orphanDocs
-  logDebug $ T.unwords ["Found", T.pack . show $ orphanCount, "orphans"]
+  orphanDocs <- runQuery Q.orphansM
+  logDebug $ T.unwords ["Found", T.pack . show . length $ orphanDocs, "orphans"]
   return $ entityVal <$> orphanDocs
 
-runUnreachable :: App [Document]
+runUnreachable :: (Monad m, RunsQuery m, Logs m) => m [Document]
 runUnreachable = do
-  connString <- asks $ dbConnString . config
-  orphanDocs <- Q.unreachable connString
-  let orphanCount = length orphanDocs
-  logDebug $ T.unwords ["Found", T.pack . show $ orphanCount, "unreachable"]
+  orphanDocs <- runQuery Q.unreachableM
+  logDebug
+    $ T.unwords ["Found", T.pack . show . length $ orphanDocs, "unreachable"]
   return $ entityVal <$> orphanDocs
 
-runNonexistant :: App [Edge]
+runNonexistant :: (Monad m, RunsQuery m, Logs m) => m [Edge]
 runNonexistant = do
-  connString <- asks $ dbConnString . config
-  nonexes    <- Q.nonexistant connString
-  let count = length nonexes
-  logDebug $ T.unwords ["Found", T.pack . show $ count, "nonexistant"]
+  nonexes <- runQuery Q.nonexistant
+  logDebug
+    $ T.unwords ["Found", T.pack . show . length $ nonexes, "nonexistant"]
   return $ entityVal <$> nonexes
 
-runSubgraph :: SubgraphOptions -> App [FilePath]
+runSubgraph
+  :: (RunsQuery m, Logs m, Files m, HasConfig m)
+  => SubgraphOptions
+  -> m [FilePath]
 runSubgraph options@SubgraphOptions { sgTargets, sgDepth } = do
   logInfo . T.unwords $ ["Finding subgraphs"]
-  -- let protoResults = runSubgraphOnArg <$> sgTargets
-  paths <- F.foldrM (flip $ runSubgraphOnArg Q.forwardLinks sgDepth)
-                    S.empty
-                    sgTargets
+  paths <- F.foldrM
+    (flip $ runSubgraphOnArg (runQuery . Q.forwardLinks) sgDepth)
+    S.empty
+    sgTargets
   return $ S.toList paths
 
-type LinkGetter = T.Text -> FilePath -> App [Entity Document]
+type LinkGetter m = FilePath -> m [Entity Document]
+
 runSubgraphOnArg
-  :: LinkGetter
+  :: (Monad m, HasConfig m, Files m)
+  => LinkGetter m
   -> Integer
   -> HashSet FilePath
   -> SubgraphTarget
-  -> App (HashSet FilePath)
+  -> m (HashSet FilePath)
 runSubgraphOnArg linkGetter maxDepth foundPaths (FileTarget path) = do
-  libPath <- asks $ libraryPath . config
-  absPath <- liftIO $ trueAbsolutePath path
+  libPath <- libraryPath <$> getConfig
+  absPath <- trueAbsolutePath path
   let relPath = makeRelative libPath absPath
   -- TODO: fix infinite depth to be a real value instead of this hack
   runSubgraphPath' linkGetter maxDepth 0 foundPaths relPath
 
-runSubgraphOnArg _ _ _ _ = throwError "Tag subgraph not yet implemented"
+runSubgraphOnArg _ _ _ _ = pure S.empty -- TODO
 
 runSubgraphPath'
-  :: LinkGetter
+  :: (Monad m)
+  => LinkGetter m
   -> Integer
   -> Integer
   -> S.HashSet FilePath
   -> FilePath
-  -> App (HashSet FilePath)
+  -> m (HashSet FilePath)
 runSubgraphPath' linkGetter maxDepth currentDepth foundPaths newPath = do
   let alreadyExists = S.member newPath foundPaths
       pastMaxDepth  = currentDepth == maxDepth
   if alreadyExists || pastMaxDepth
     then pure foundPaths
     else do
-      connString <- asks $ dbConnString . config
-      children   <- linkGetter connString newPath
+      children <- linkGetter newPath
       let setWithCurrent = S.insert newPath foundPaths
       let childPaths     = documentPath . entityVal <$> children
       F.foldrM
@@ -119,11 +126,11 @@ runSubgraphPath' linkGetter maxDepth currentDepth foundPaths newPath = do
         setWithCurrent
         childPaths
 
--- not working, not sure why. it's only returning the child
 runBacklinks :: BacklinkOptions -> App [FilePath]
 runBacklinks options@BacklinkOptions { blTargets, blDepth } = do
   logInfo . T.unwords $ ["Finding backlinks"]
-  paths <- F.foldrM (flip $ runSubgraphOnArg Q.backwardLinks blDepth)
-                    S.empty
-                    blTargets
+  paths <- F.foldrM
+    (flip $ runSubgraphOnArg (runQuery . Q.backwardLinks) blDepth)
+    S.empty
+    blTargets
   return $ S.toList paths
